created new package for runners removed some annotations, renamed class
