+++ src/main/java/pl/marcin/project/ProjectApplication.java
@@ -1,17 +1,13 @@
 package pl.marcin.project;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication
-
 public class ProjectApplication {
 
     public static void main(String[] args) {
+
         SpringApplication.run(ProjectApplication.class, args);
     }
 }
