+++ src/main/java/pl/marcin/project/RunnerFactory.java
@@ -1,23 +1,24 @@
 package pl.marcin.project;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Configuration;
+import pl.marcin.project.runners.AppRunner;
+import pl.marcin.project.runners.ShopAppRunner;
+import pl.marcin.project.runners.WebAppRunner;
 
-@Configuration
-@RequiredArgsConstructor
-public class AppRunnerConfig implements ApplicationRunner {
+
+public class RunnerFactory implements ApplicationRunner {
 
     private AppRunner appRunner;
 
     @Override
     public void run(ApplicationArguments args) throws Exception {
         String runningWay = "shop";
-        if (runningWay.equals("web")) {
+        if ("web".equals(runningWay)) {
             appRunner = new WebAppRunner();
-        } else if (runningWay.equals("shop")) {
+        } else if ("shop".equals(runningWay)) {
             appRunner = new ShopAppRunner();
         }
         appRunner.runApplication();
