+++ src/main/java/pl/marcin/project/runners/AppRunner.java
@@ -1,4 +1,4 @@
-package pl.marcin.project;
+package pl.marcin.project.runners;
 
 public interface AppRunner {
     void runApplication();
