+++ src/main/java/pl/marcin/project/runners/WebAppRunner.java
@@ -1,8 +1,11 @@
-package pl.marcin.project;
+package pl.marcin.project.runners;
+
+import pl.marcin.project.runners.AppRunner;
 
 public class WebAppRunner implements AppRunner {
+
     @Override
     public void runApplication() {
-        System.out.println("Web is running...");
+
     }
 }
